vulnerability root
